+++ scripts/launch_in_docker.sh
@@ -13,8 +13,5 @@ else
 	npm run db:init
 fi
 
-# Deploy commands
-node . --deploy
-
 # Launch
 node .